Prevent enemy from beating a dead player.
I'm not super sure what I want the enemy to do when it
wins. I like the idea of other adventurers being able to
find the dead adventurer and rez them with a special
potion, so maybe they have to fight off the enemy when
they get there before they can rez. Or maybe the enemy
should just run off at the end of the battle. Maybe it
should depend on the type of enemy?

Either way I'm a dope and attached the game logic to the
view panel, so we've got to go through the process of
decoupling that before we can even think about adventurers
interacting.